refactor: pipe 파일명 변경
