+++ src/index.js
@@ -2,7 +2,7 @@ import { getCarNames, getMoveCount } from "./view/input.js";
 import playGame from "./domain/playGame.js";
 import findWinners from "./domain/findWinners.js";
 import { displayWinner } from "./view/output.js";
-import { pipe } from "./utils/functional.js";
+import { pipe } from "./utils/pipe.js";
 
 async function run() {
   const carNames = await getCarNames();
